fix(autoware_euclidean_cluster_object_detector): fix bugprone-narrowing-conversions warnings (#938)
* fix(autoware_euclidean_cluster_object_detector): fix bugprone-narrowing-conversions warnings


* style(pre-commit): autofix

---------

+++ perception/autoware_euclidean_cluster_object_detector/lib/voxel_grid_based_euclidean_cluster.cpp
@@ -99,7 +99,7 @@ bool VoxelGridBasedEuclideanCluster::cluster(
 
   // create voxel
   pcl::PointCloud<pcl::PointXYZ>::Ptr pointcloud(new pcl::PointCloud<pcl::PointXYZ>);
-  int point_step = pointcloud_msg->point_step;
+  int point_step = static_cast<int>(pointcloud_msg->point_step);
   pcl::fromROSMsg(*pointcloud_msg, *pointcloud);
   pcl::PointCloud<pcl::PointXYZ>::Ptr voxel_map_ptr(new pcl::PointCloud<pcl::PointXYZ>);
   voxel_grid_.setLeafSize(voxel_leaf_size_, voxel_leaf_size_, 100000.0);
@@ -142,7 +142,7 @@ bool VoxelGridBasedEuclideanCluster::cluster(
     const auto & cluster = cluster_indices.at(cluster_idx);
     auto & temporary_cluster = temporary_clusters.at(cluster_idx);
     for (const auto & point_idx : cluster.indices) {
-      map[point_idx] = cluster_idx;
+      map[point_idx] = static_cast<int>(cluster_idx);
     }
     temporary_cluster.height = pointcloud_msg->height;
     temporary_cluster.fields = pointcloud_msg->fields;
